Move dimension mappers up into containers
